Change ANNA dataset

+++ HideWordSolver/Solver/research_word.c
@@ -2,8 +2,8 @@
 
 #define research research_word_with_direction
 
-char research_word(const char grid[MAX_SIZE][MAX_SIZE], size_t grid_size_x,
-	size_t grid_size_y, const char *word, size_t word_size,
+char research_word(const char grid[MAX_SIZE][MAX_SIZE], int grid_size_x,
+	int grid_size_y, const char *word, int word_size,
 	int word_position[4])
 {
 	size_t i = 0;
